+++ packages/frontend/src/scripts/sims/healer/ast_sheet_sim.ts
@@ -1,7 +1,7 @@
-import {BuffSettingsExport} from "../party_comp_settings";
 import {GcdAbility, OgcdAbility, SimSettings, SimSpec} from "@xivgear/core/sims/sim_types";
 import {CycleProcessor, CycleSimResult, ExternalCycleSettings, Rotation} from "@xivgear/core/sims/cycle_sim";
 import {BaseMultiCycleSim} from "../sim_processors";
+import {BuffSettingsExport} from "@xivgear/core/sims/common/party_comp_settings";
 
 
 const filler: GcdAbility = {
